% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsm_plot.R
\name{plot.fsm_od_totals}
\alias{plot.fsm_od_totals}
\title{Plot OD Totals}
\usage{
\method{plot}{fsm_od_totals}(x, ...)
}
\arguments{
\item{x}{An \code{fsm_od_totals} object.}

\item{...}{Additional arguments passed to \code{graphics::barplot()}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Plots observed productions and attractions from an
\code{fsm_od_totals} object.
}
\examples{
od <- fsm_toy_od
totals <- fsm_od_totals(od)
plot(totals)
}
