% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsm_inspect.R
\name{print.fsm_generation}
\alias{print.fsm_generation}
\title{Print a Generation Model}
\usage{
\method{print}{fsm_generation}(x, ...)
}
\arguments{
\item{x}{An \code{fsm_generation} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Generation Model
}
