% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsm_inspect.R
\name{print.fsm_od}
\alias{print.fsm_od}
\title{Print an Origin-Destination Object}
\usage{
\method{print}{fsm_od}(x, ..., n = 6L)
}
\arguments{
\item{x}{An \code{fsm_od} object.}

\item{...}{Additional arguments passed to the underlying data.table print
method.}

\item{n}{Maximum number of OD pairs to display. Use \code{Inf} to display
every pair.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an Origin-Destination Object
}
