% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsm_inspect.R
\name{print.fsm_od_totals}
\alias{print.fsm_od_totals}
\title{Print Production and Attraction Totals}
\usage{
\method{print}{fsm_od_totals}(x, ..., n = 6L)
}
\arguments{
\item{x}{An \code{fsm_od_totals} object.}

\item{...}{Additional arguments passed to the underlying data.table print
method.}

\item{n}{Maximum number of rows to display. Use \code{Inf} to display every
row.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Production and Attraction Totals
}
