% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsm_inspect.R
\name{print.fsm_population}
\alias{print.fsm_population}
\title{Print a Population Attribute Object}
\usage{
\method{print}{fsm_population}(x, ..., n = 6L)
}
\arguments{
\item{x}{An \code{fsm_population} object.}

\item{...}{Additional arguments passed to the underlying data.table print
method.}

\item{n}{Maximum number of disaggregate records to display. Use \code{Inf} to
display every record.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Population Attribute Object
}
