% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsm_inspect.R
\name{print.fsm_trip}
\alias{print.fsm_trip}
\title{Print an Origin-Destination Survey Object}
\usage{
\method{print}{fsm_trip}(x, ..., n = 6L)
}
\arguments{
\item{x}{An \code{fsm_trip} object.}

\item{...}{Additional arguments passed to the underlying data.table print method.}

\item{n}{Maximum number of trip records to display. Use \code{Inf} for all records.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an Origin-Destination Survey Object
}
