% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsm_inspect.R
\name{print.fsm_zone}
\alias{print.fsm_zone}
\title{Print a Zone Attribute Object}
\usage{
\method{print}{fsm_zone}(x, ..., n = 6L)
}
\arguments{
\item{x}{An \code{fsm_zone} object.}

\item{...}{Additional arguments passed to the underlying data.table print
method.}

\item{n}{Maximum number of zones to display. Use \code{Inf} to display every
zone.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Zone Attribute Object
}
