% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsm_inspect.R
\name{print.summary.fsm_od_totals}
\alias{print.summary.fsm_od_totals}
\title{Print a Production and Attraction Totals Summary}
\usage{
\method{print}{summary.fsm_od_totals}(x, ...)
}
\arguments{
\item{x}{A \code{summary.fsm_od_totals} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Production and Attraction Totals Summary
}
