% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsm_inspect.R
\name{print.summary.fsm_population}
\alias{print.summary.fsm_population}
\title{Print a Population Attribute Summary}
\usage{
\method{print}{summary.fsm_population}(x, ...)
}
\arguments{
\item{x}{A \code{summary.fsm_population} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Population Attribute Summary
}
