% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsm_inspect.R
\name{print.summary.fsm_trip}
\alias{print.summary.fsm_trip}
\title{Print an Origin-Destination Survey Summary}
\usage{
\method{print}{summary.fsm_trip}(x, ...)
}
\arguments{
\item{x}{A \code{summary.fsm_trip} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an Origin-Destination Survey Summary
}
