% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsm_inspect.R
\name{summary.fsm_od}
\alias{summary.fsm_od}
\title{Summarize an Origin-Destination Object}
\usage{
\method{summary}{fsm_od}(object, ...)
}
\arguments{
\item{object}{An \code{fsm_od} object.}

\item{...}{Unused.}
}
\value{
A list with core OD diagnostics.
}
\description{
Summarize an Origin-Destination Object
}
